# services/main-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.MeResponse:
    properties:
      email:
        type: string
      emailVerifiedAt:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      preferences:
        $ref: '#/definitions/handler.UserPreferencesResponse'
    type: object
...
  handler.UpdateMeRequest:
    properties:
      brandOpenness:
        type: string
      budgetSensitivity:
        type: string
      defaultCurrency:
        type: string
      name:
        type: string
      phone:
        type: string
      preferredCategories:
        items:
          type: string
        type: array
    type: object
...
  handler.UserPreferencesResponse:
    properties:
      brandOpenness:
        type: string
      budgetSensitivity:
        type: string
      defaultCurrency:
        type: string
      preferredCategories:
        items:
          type: string
        type: array
    type: object
...
  internal_users_handler.ErrorResponse:
    properties:
      code:
        example: VALIDATION_FAILED
        type: string
      detail:
        example: email and name are required
        type: string
      instance:
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
      type:
        example: about:blank
        type: string
    type: object
...
host: localhost:18080
info:
  contact: {}
  description: Modular monolith REST API for React web and native clients.
  title: Server API
  version: "1.0"
paths:
...
  /users/me:
    delete:
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_users_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_users_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_users_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete account and all user data (FR-037)
      tags:
      - users
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_users_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      parameters:
      - description: Profile update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_users_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update profile and preferences
      tags:
      - users
schemes:
- http
swagger: "2.0"